Attempt to fix weird numbers

+++ src/share/classes/com/sun/btrace/aggregation/Minimum.java
@@ -36,7 +36,7 @@ class Minimum implements AggregationValue {
 
     @Override
     public synchronized void clear() {
-        min = Integer.MAX_VALUE;
+        min = Long.MAX_VALUE;
     }
 
     @Override
